requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
psycopg2-binary
python-dotenv
sqlmodel